+++ qiskit-superstaq/qiskit_superstaq/superstaq_backend.py
@@ -423,3 +423,54 @@ class SuperstaqBackend(qiskit.providers.BackendV1):
         if circuits_is_list:
             return resource_estimates
         return resource_estimates[0]
+
+    def submit_aces(
+        self,
+        qubits: qiskit.QuantumRegister,
+        shots: int,
+        num_circuits: int,
+        mirror_depth: int,
+        extra_depth: int,
+        **kwargs: Any,
+    ) -> str:
+        """Submits the jobs to characterize this target through the ACES protocol.
+
+        Args:
+            qubits: Register of qubits to characterize.
+            shots: How many shots to use per circuit submitted.
+            num_circuits: How many random circuits to use in the protocol.
+            mirror_depth: The half-depth of the mirror portion of the random circuits.
+            extra_depth: The depth of the fully random portion of the random circuits.
+            kwargs: Other execution parameters.
+                - tag: Tag for all jobs submitted for this protocol.
+                - lifespan: How long to store the jobs submitted for in days (only works with right
+                permissions).
+                - method: Which type of method to execute the circuits with.
+
+        Returns:
+            A string with the job id for the ACES job created.
+
+        Raises:
+            ValueError: If any the target passed are not valid.
+            SuperstaqServerException: if the request fails.
+        """
+        return self._provider._client.submit_aces(
+            target=self.name(),
+            qubits=[q._index for q in qubits],
+            shots=shots,
+            num_circuits=num_circuits,
+            mirror_depth=mirror_depth,
+            extra_depth=extra_depth,
+            **kwargs,
+        )
+
+    def process_aces(self, job_id: str) -> List[float]:
+        """Process the jobs submitted by `submit_aces` and get the gate eigenvalues.
+
+        Args:
+            job_id: The job id returned by `submit_aces`.
+
+        Returns:
+            The estimated eigenvalues.
+        """
+        return self._provider._client.process_aces(job_id=job_id)
